backend/requirements.txt
django==5.2
djangorestframework==3.15.2
psycopg[binary]>=3.1.8
celery==5.3.6
redis==5.0.3
django-cors-headers==4.3.1
pytest==8.3.5
pytest-django==4.8.0
pytest-xdist==3.6.1
faker==24.11.0
gunicorn==21.2.0
honcho==1.1.0
dj-database-url==2.1.0
whitenoise==6.6.0
